+++ CHANGELOG.md
@@ -51,6 +51,8 @@ a release.
 - `Gedmo\Tool\Logging\DBAL\QueryAnalizer` class without replacement.
 - Using YAML mapping is deprecated, you SHOULD migrate to attributes, annotations or XML.
 - `Gedmo\Mapping\Event\AdapterInterface::__call()` method.
+- `Gedmo\Tool\Wrapper\AbstractWrapper::clear()` method.
+- `Gedmo\Tool\Wrapper\WrapperInterface::populate()` method.
 
 ### Changed
 - In order to use a custom cache for storing configuration of an extension, the user has to call `setCacheItemPool()`